+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NGSI Go v0.9.0-next
 
+-   Hardening: Add feature to support OpenID for Keyrock (#206)
 -   Update: Update docker images (#205)
 -   Improve: Add --serial flag to print serial value as NGSI Go version (#204)
 -   Improve: Add colon as a valid character for alias name (#203)